Split tests for dataclasses, final classes, attrs and pydantic.
